+++ app/components/ProcessesGrid/ProcessesGrid.php
@@ -173,6 +173,21 @@ class ProcessesGrid extends GridBuilder implements IGridExtendingComponent {
 
             return $el;
         };
+
+        $deleteInstance = $this->addAction('deleteInstance');
+        $deleteInstance->setTitle('Delete instance');
+        $deleteInstance->onCanRender[] = function(DatabaseRow $row, Row $_row, Action &$action) {
+            return $this->processAuthorizator->canUserDeleteProcessInstance($row->instanceId, $this->app->currentUser->getId());
+        };
+        $deleteInstance->onRender[] = function(mixed $primaryKey, DatabaseRow $row, Row $_row, HTML $html) {
+            $el = HTML::el('a')
+                ->class('grid-link')
+                ->href($this->createFullURLString('User:Processes', 'deleteInstance', ['instanceId' => $primaryKey, 'processId' => $row->processId, 'view' => $this->view]))
+                ->text('Delete instance')
+            ;
+
+            return $el;
+        };
     }
 
     /**