+++ app/modules/AGUICore.php
@@ -114,13 +114,6 @@ abstract class AGUICore {
         $components = TemplateHelper::loadComponentsFromTemplateContent($template->getTemplateContent());
 
         foreach($components as $componentName => $componentAction) {
-            /**
-             * If this is an ajax request and the query parameter "do" is not null, then only the requested component should be called and processed because other components are not needed
-             */
-            /*if($this->httpRequest->get('do') !== null && $this->httpRequest->isAjax) {
-                $doComponentName = explode('-', $this->httpRequest->get('do'))[0];
-                if($doComponentName != $componentName) continue;
-            }*/
             if(method_exists($this, $componentAction)) {
                 if(isset($_GET['isFormSubmit']) && $_GET['isFormSubmit'] == '1') {
                     $fr = $this->createFormRequest();