+++ app/modules/UserModule/Presenters/ProcessesPresenter.php
@@ -5,7 +5,9 @@ namespace App\Modules\UserModule;
 use App\Components\ProcessesGrid\ProcessesGrid;
 use App\Components\ProcessViewsSidebar\ProcessViewsSidebar;
 use App\Constants\Container\ProcessGridViews;
+use App\Constants\JobQueueTypes;
 use App\Core\Http\HttpRequest;
+use App\Exceptions\AException;
 
 class ProcessesPresenter extends AUserPresenter {
     public function __construct() {
@@ -62,6 +64,29 @@ class ProcessesPresenter extends AUserPresenter {
         );
 
         return $processGrid;
+    }
+
+    public function handleDeleteInstance() {
+        $instanceId = $this->httpRequest->get('instanceId');
+        $view = $this->httpRequest->get('view');
+
+        try {
+            $this->app->jobQueueRepository->beginTransaction(__METHOD__);
+
+            $this->app->jobQueueManager->insertNewJob(JobQueueTypes::DELETE_CONTAINER_PROCESS_INSTANCE, [
+                'instanceId' => $instanceId
+            ], null);
+
+            $this->app->jobQueueRepository->commit($this->getUserId(), __METHOD__);
+
+            $this->flashMessage('Process instance was enqueued for deletion.', 'success');
+        } catch(AException $e) {
+            $this->app->jobQueueRepository->rollback(__METHOD__);
+
+            $this->flashMessage('Could not enqueue process instance for deletion. Reason: ' . $e->getMessage(), 'error', 10);
+        }
+
+        $this->redirect($this->createURL('list', ['view' => $view]));
     }
 }