+++ data/db/migrations/migration_2025-06-16_0015_job_queue_processing_history.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Data\Db\Migrations;
+
+use App\Core\DB\ABaseMigration;
+use App\Core\DB\Helpers\TableSchema;
+use App\Core\DB\Helpers\TableSeeding;
+
+class migration_2025_06_16_0015_job_queue_processing_history extends ABaseMigration {
+    public function up(): TableSchema {
+        $table = $this->getTableSchema();
+
+        $table->create('job_queue_processing_history')
+            ->primaryKey('entryId')
+            ->foreignKey('jobId', true)
+            ->text('description')
+            ->enum('type')
+            ->datetimeAuto('dateCreated');
+
+        return $table;
+    }
+
+    public function down(): TableSchema {
+        return $this->getTableSchema();
+    }
+
+    public function seeding(): TableSeeding {
+        return $this->getTableSeeding();
+    }
+}
+
+?>