+++ services/cofrs_master.php
@@ -8,7 +8,7 @@ require_once('CommonService.php');
 global $app;
 
 try {
-    $service = new ContainerOrphanedFilesRemovingMasterService($app->logger, $app->serviceManager, $app->containerManager);
+    $service = new ContainerOrphanedFilesRemovingMasterService($app);
     $service->run();
 } catch(Exception|Error $e) {
     throw new ServiceException($e->getMessage(), $e);