+++ services/job_queue_service.php
@@ -1,14 +1,14 @@
 <?php
 
 use App\Exceptions\ServiceException;
-use App\Services\LogRotateService;
+use App\Services\JobQueueService;
 
 require_once('CommonService.php');
 
 global $app;
 
 try {
-    $service = new LogRotateService($app);
+    $service = new JobQueueService($app);
     $service->run();
 } catch(Exception|Error $e) {
     throw new ServiceException($e->getMessage(), $e);